convert test_change_pools to WalletTestFramework
